# specs/SPEC/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Luminary Lab API",
    "version": "1.0.0",
    "description": "API for Luminary Lab - AI-powered professional photo retouching",
    "contact": {
      "name": "Luminary Lab Engineering",
      "email": "engineering@luminarylabs.com"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    },
    {
      "url": "https://api.luminarylabs.com/api/v1",
      "description": "Production API"
    }
  ],
...
  "paths": {
...
    "/projects/{projectId}": {
      "get": {
        "tags": ["projects"],
        "summary": "Get project details",
        "operationId": "getProject",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectWithImages"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": ["projects"],
        "summary": "Update project",
        "operationId": "updateProject",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              },
              "example": {
                "name": "Updated Project Name"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["projects"],
        "summary": "Delete project",
        "operationId": "deleteProject",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Project deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/admin/presets": {
      "post": {
        "tags": ["admin"],
        "summary": "Create new preset (admin only)",
        "operationId": "createPreset",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePresetRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Preset created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Preset"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/admin/presets/{presetId}": {
      "put": {
        "tags": ["admin"],
        "summary": "Update preset (admin only)",
        "operationId": "updatePreset",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "presetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePresetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preset updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Preset"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": ["admin"],
        "summary": "Delete preset (admin only)",
        "operationId": "deletePreset",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "presetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Preset deleted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "UpdateProjectRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200
          }
        }
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProjectWithImages": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Project"
          },
          {
            "type": "object",
            "properties": {
              "originalImage": {
                "$ref": "#/components/schemas/Image"
              },
              "processedImages": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ProcessedImage"
                }
              }
            }
          }
        ]
      },
...
      "Image": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "fileFormat": {
            "type": "string"
          },
          "fileSize": {
            "type": "integer"
          },
          "width": {
            "type": "integer"
          },
          "height": {
            "type": "integer"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProcessedImage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "format": {
            "type": "string"
          },
          "colorSpace": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "downloadUrl": {
            "type": "string",
            "description": "Temporary download URL"
          }
        }
      },
      "Preset": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "previewUrl": {
            "type": "string",
            "description": "Before/after preview URL"
          }
        }
      },
...
      "CreatePresetRequest": {
        "type": "object",
        "required": ["name", "description", "promptTemplate", "blendingParameters", "thumbnailUrl"],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string"
          },
          "promptTemplate": {
            "type": "string"
          },
          "blendingParameters": {
            "type": "object"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "previewUrl": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "UpdatePresetRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "description": {
            "type": "string"
          },
          "promptTemplate": {
            "type": "string"
          },
          "blendingParameters": {
            "type": "object"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "previewUrl": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error code"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          },
          "details": {
            "type": "object",
            "description": "Additional error details"
          }
        }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "UNAUTHORIZED",
              "message": "Authentication required"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "FORBIDDEN",
              "message": "Insufficient permissions"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "NOT_FOUND",
              "message": "Resource not found"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token authentication"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}